% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{check_all_events}
\alias{check_all_events}
\title{Check that all unique event names exist in linked arms}
\usage{
check_all_events(data, linked_arms)
}
\arguments{
\item{data}{a dataframe}

\item{linked_arms}{a dataframe output from [link_arms_rsd()] data}
}
\description{
Check that all unique event names exist in linked arms
}
\keyword{internal}
