% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{join_linked_arms}
\alias{join_linked_arms}
\title{Join the a linked arms dataset onto a supertibble data tibble}
\usage{
join_linked_arms(data_tbl, linked_arms)
}
\arguments{
\item{data_tbl}{a data tibble from a supertibble}

\item{linked_arms}{linked arms data}
}
\value{
a data tibble appended with linked arms data
}
\description{
Using linked arms data, join onto each data tibble to make
the original event name accessible. During \code{\link[REDCapTidieR:read_redcap]{REDCapTidieR::read_redcap()}} operations,
event names have an "arm" suffix that gets dropped. This looks to reinstate it
for proper links later on in the \code{\link[=record_status_dashboard]{record_status_dashboard()}} logic.
}
\keyword{internal}
