% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{link_arms_rsd}
\alias{link_arms_rsd}
\title{Link longitudinal REDCap instruments with their events/arms}
\usage{
link_arms_rsd(redcap_uri, token, suppress_redcapr_messages = TRUE)
}
\arguments{
\item{redcap_uri}{The REDCap URI}

\item{token}{The REDCap API token}

\item{suppress_redcapr_messages}{A logical to control whether to suppress messages
from REDCapR API calls. Default `TRUE`.}
}
\value{
Returns a \code{tibble} of \code{redcap_event_name}s with list elements
containing a vector of associated instruments.
}
\description{
For REDCap databases containing arms and events, it is necessary to determine
how these are linked and what variables belong to them.
}
\details{
This function is built similar to, but varies from [link_arms()] to reduce
API calls and check failed status messages as an indicator for whether or not
a database `has_arms`.
}
\keyword{internal}
