% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{prepare_supertbl_data}
\alias{prepare_supertbl_data}
\title{Prepare supertibble and tidy tibbles for record status dashboard processing}
\usage{
prepare_supertbl_data(redcap_uri, token)
}
\arguments{
\item{redcap_uri}{The
URI/URL of the REDCap server (e.g.,
"https://server.org/apps/redcap/api/"). Required.}

\item{token}{The user-specific string that serves as the password for a
project. Required.}
}
\value{
a list
}
\description{
Helper function that returns the supertibble and tidy tibbles
from it.
}
\keyword{internal}
