% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{prepare_tidy_tbls}
\alias{prepare_tidy_tbls}
\title{Prepare supertibble and tidy tibbles for record status dashboard processing}
\usage{
prepare_tidy_tbls(supertbl)
}
\value{
a list
}
\description{
Helper function that returns the supertibble and tidy tibbles
from it.
}
\keyword{internal}
