% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{combine_data}
\alias{combine_data}
\title{Join supertibble datatibbles and arrange by instrument/event order}
\usage{
combine_data(
  data,
  record_id_field,
  instrument_order,
  event_order,
  has_arms,
  linked_arms = NULL
)
}
\arguments{
\item{data}{a named list of data tibbles from the supertbl}

\item{record_id_field}{the record ID field for the REDCap project identified by
[REDCapTidieR:::get_record_id_field()]}

\item{instrument_order}{a vector determining the factor level order of the project instruments}

\item{event_order}{a vector determining the factor level order of the project events}

\item{has_arms}{TRUE/FALSE whether or not the project is longitudinal/has arms}

\item{linked_arms}{if a longitudinal project, use the linked_arms data output from
[link_arms_rsd()]. Default `NULL`.}
}
\value{
a dataframe
}
\description{
This function takes a list of data tibbles and binds together
the columns necessary for the record status dashboard view:

- the record ID
- the form completion status
- the unique event name
- the redcap form label
- the redcap event label (if longitudinal)
}
\keyword{internal}
