% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_status_dashboard.R
\name{record_status_dashboard}
\alias{record_status_dashboard}
\title{Create a Record Status Dashboard Dataframe}
\usage{
record_status_dashboard(supertbl)
}
\value{
A dataframe representing the Record Status Dashboard view of a REDCap project.
Each row corresponds to a record, and columns represent different forms, events, and their
respective status indicators.
}
\description{
[record_status_dashboard()] creates a dataframe that replicates the display of
the Record Status Dashboard from a REDCap project, given a REDCap API URI and token.
This function provides a structured overview of the status of records and their
associated instruments and events, similar to what is seen in the REDCap web interface.

This output can be utilized for visualizations, such as heatmap displays, to gain
insights into data completeness and record status trends.
}
\details{
The REDCap Record Status Dashboard is a widely used tool for navigating and
understanding the state of records within a REDCap project, displaying their
completion statuses across various instruments and events. It can show whether
forms are marked as complete, in progress/unverified, incomplete, or not yet opened.

[record_status_dashboard()] captures this information and provides it in a
structured dataframe format. By default, the function converts form completion
status into a percentage format, accounting for repeating instruments with overlapping
form status fields. This allows for a more flexible representation of the data,
especially useful when dealing with projects that utilize repeating forms or
multiple events.
}
\examples{
\dontrun{
# Example usage:
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")
record_status_dashboard(
  redcap_uri,
  token
)
}
}
