% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapExploreR-package.R
\docType{package}
\name{REDCapExploreR-package}
\alias{REDCapExploreR}
\alias{REDCapExploreR-package}
\title{REDCapExploreR: Exploratory Data Quality Tools for REDCap Projects}
\description{
Provides tools to explore REDCap projects, summarize project structure, and flag common data quality issues for data managers.
}
\keyword{internal}
