% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapExploreR-package.R
\docType{package}
\name{REDCapExploreR-package}
\alias{REDCapExploreR}
\alias{REDCapExploreR-package}
\title{REDCapExploreR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
